+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Replace bootstrap NPM dependency with a package dependency on pwix:ui-ui-bootstrap5
     - Add (missing) tracker dependency
     - Obsolete Roles.removeUserAssignmentsForRoles() function, replaced with Roles.removeUserAssignmentsFromRoles()
+    - Obsolete Roles.removeAllRolesFromUser() function, replaced with Roles.removeAssignedRolesFromUser()
 
 ### 1.3.1